Bug Fix, Maintain correctness among optimization flags
shld, shrd write to memory but did not inform compiler to clobber it.
Clobber "memory" does not help when compiling with -O2 on x64 linux, but
it works on mac (gcc 9.2)
We fix it by storing results to a local variable and write it to
desinated memory location in C code.
Now we can compile abn with -O2 flags.